+++ packages/popmotion-react/stories/TransitionGroup.js
@@ -1,0 +1,65 @@
+import React from 'react';
+import TransitionGroup from 'react-transition-group/TransitionGroup';
+import { MotionValue } from '../src';
+import { tween, transform } from 'popmotion';
+import styled from 'styled-components';
+const { px } = transform;
+
+const Box = styled.div`
+  height: 100px;
+  background: red;
+`;
+
+const stateChangeHandlers = {
+  componentWillEnter: ({ value, onComplete }) => tween({
+    from: value.get(),
+    to: 100,
+    onUpdate: value,
+    onComplete
+  }).start(),
+  componentWillLeave: ({ value, onComplete }) => tween({
+    from: value.get(),
+    to: 0,
+    onUpdate: value,
+    onComplete
+  }).start()
+};
+
+
+export default class TransitionGroupExample extends React.Component {
+  state = {
+    numChildren: 0
+  };
+
+  componentDidMount() {
+    this.interval = setInterval(() => {
+      const { numChildren } = this.state;
+      this.setState({
+        numChildren: numChildren + 1
+      });
+    }, 1000);
+  }
+
+  componentWillUnmount() {
+    clearInterval(this.interval);
+  }
+
+  render() {
+    const children = [];
+    const { numChildren } = this.state;
+
+    for (let i = 0; i < numChildren; i++) {
+      children.push(
+        <MotionValue key={i} onStateChange={stateChangeHandlers}>
+          {({ v }) => <Box style={{ width: px(v) }} />}
+        </MotionValue>
+      );
+    }
+
+    return (
+      <TransitionGroup>
+        {children}
+      </TransitionGroup>
+    );
+  }
+}