+++ packages/popmotion-react/stories/index.js
@@ -5,6 +5,7 @@ import { action } from '@storybook/addon-actions';
 import { linkTo } from '@storybook/addon-links';
 
 import DragSingleChild from './DragSingleChild';
+import TransitionGroup from './TransitionGroup';
 import Toggle from './Toggle';
 
 // storiesOf('Welcome', module).add('to Storybook', () => <Welcome showApp={linkTo('Button')} />);
@@ -15,6 +16,7 @@ import Toggle from './Toggle';
 
 
 storiesOf('MotionValue').add('Drag single child', () => <DragSingleChild />);
+storiesOf('MotionValue').add('React TransitionGroup', () => <TransitionGroup />);
 //storiesOf('MotionValue').add('Rotate Three.js Cube', () => <RotateThreeCube />);
 storiesOf('MotionValue').add('Toggle state', () => <Toggle />);